ref(metrics): refactor metrics pkg for derived metrics
