+++ src/sentry/snuba/metrics/__init__.py
@@ -1,2 +1,4 @@
 from .datasource import *  # NOQA
-from .helpers import *  # NOQA
+from .fields import *  # NOQA
+from .query_builder import *  # NOQA
+from .utils import *  # NOQA
